* Add an input for Gist Name, prepopulated from default.
* Remove "/140" from earlier commit.

I had to include the session variable in the server, since Shiny needs that to updateInput(). I put the filename update inside the tweet button handler since triggering of text update is irratic (get partial filename updates) and I didn't want to add another button to the UI. It seems to work fine.